Bugfix: Disappearing cursor
`RoR2.ColorCatalog.multiplayerColors` is readonly. The compiler seems
to make optimizations when compiling the function `RoR2.ColorCatalog.GetMultiplayerColor()`
inlining the reference to `RoR2.ColorCatalog.multiplayerColors`. When
the value was set in HookManager.UpdateMultiplayerColors(), the
reference was changed and made the reference in
`RoR2.ColorCatalog.GetMultiplayerColor()' invalid.

To solve this, I store the multiplayer colors in this mod instead and
patch `RoR2.ColorCatalog.GetMultiplayerColor()' to use that instead.

Looking at the dlls, it seems the only code that uses MultiplayerColors
in the RoR2 managed is for the cursor so I don't forsee any issues with
moving this over.